renormalize
Bounding box is now [0,1] for any target input.
Reg energies are mean per point, which combined with the bounding box make them united.
Fit energy is normalized by the it's value at initialization meaning it's 1 at the start.
The normalization factor for fit energy is 70 - 90 for the non-rescaled bounding box and is .5 - 1.5 for the normalized box.
Regularizer energies were summed over 3000 points but is now the mean meaning it's dropped by a factor of 3000.

This means any scaling I put on regularizers pre-normalization need to be increased 40x